Adding Redirect Logic
If a request is now made to the API server that is not defined as a route,
our custom redirect handler will execute and send the user back to
dev.jetway.io. The idea is that if a dev is playing around with the API and
can't figure it out, we should send them to the developer support portal for
help.